+++ hotrace.h
